+++ src/modules/auth/AuthLogin/AuthLogin.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/camelcase */
 import React from 'react'
 import { useAuth0 } from '@auth0/auth0-react'
 import { useLocation } from 'react-router-dom'
@@ -14,10 +15,16 @@ interface AuthLoginProps {
 const AuthLogin: React.FC<AuthLoginProps> = ({ className }) => {
   const { loginWithRedirect } = useAuth0()
   const location = useLocation()
-  const text = location.pathname === Path.SIGN_UP ? 'Sign Up' : 'Login'
+  const isSignUp = location.pathname === Path.SIGN_UP
+  const text = isSignUp ? 'Sign Up' : 'Login'
 
   return (
-    <Button className={className} onClick={loginWithRedirect}>
+    <Button
+      className={className}
+      onClick={(): Promise<void> =>
+        loginWithRedirect({ screen_hint: isSignUp ? 'signup' : 'login' })
+      }
+    >
       <span className={styles.buttonContent}>{text}</span>
       <span className={styles.buttonContent}>🔑</span>
     </Button>
